Add color and intensity controls for lampadine
Introduced methods to change the color and intensity of lampadine in CasaIntelligente and GestoreLampadine. Updated Lampadina to use double for intensity, added setIntensita and improved setColore to update the drawing. Adjusted DisegnoLampadina and Main to support these changes and fixed a bug in spegnilampadina.

+++ CasaDomotica/src/DisegnoLampadina.java
@@ -62,7 +62,7 @@ public class DisegnoLampadina {
         return true;
     }
 
-    public void accendi(int intensita) {
+    public void accendi(double intensita) {
         draw(intensita);
     }
